Migrate converts to adapter namespace
