+++ SAP2000_Adapter/CRUD/Read/Bar.cs
@@ -69,10 +69,10 @@ namespace BH.Adapter.SAP2000
                     double[] springEnd = new double[6];
 
                     m_model.FrameObj.GetReleases(id, ref restraintStart, ref restraintEnd, ref springStart, ref springEnd);
-                    bhomBar.Release = Engine.SAP2000.Convert.GetBarRelease(restraintStart, springStart, restraintEnd, springEnd);
+                    bhomBar.Release = Adapter.SAP2000.Convert.GetBarRelease(restraintStart, springStart, restraintEnd, springEnd);
 
-                    //bhomBar.Release.StartRelease = Engine.SAP2000.Convert.GetConstraint6DOF(restraintStart, springStart);
-                    //bhomBar.Release.EndRelease = Engine.SAP2000.Convert.GetConstraint6DOF(restraintEnd, springEnd);
+                    //bhomBar.Release.StartRelease = Adapter.SAP2000.Convert.GetConstraint6DOF(restraintStart, springStart);
+                    //bhomBar.Release.EndRelease = Adapter.SAP2000.Convert.GetConstraint6DOF(restraintEnd, springEnd);
                     
                     string propertyName = "";
                     string sAuto = ""; //This is the name of the auto select list assigned to the frame object, if any.
