+++ SAP2000_Adapter/Convert/AdapterIdName.cs
@@ -20,28 +20,16 @@
  * along with this code. If not, see <https://www.gnu.org/licenses/lgpl-3.0.html>.      
  */
 
-using BH.oM.Geometry;
-using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
-
-namespace BH.Engine.SAP2000
+namespace BH.Adapter.SAP2000
 {
-    public static partial class Convert
-    {   
+    public partial class Convert
+    {
         /***************************************************/
-        /**** Public Methods                            ****/
-        /***************************************************/
-
-        public static Vector BarLocalAxisToCSI(this Vector axisBHoM)
-        {
-            return Geometry.Modify.Transform(axisBHoM, barLocalAxisToCSI);
-        }
-
+        /**** Public Strings                            ****/
         /***************************************************/
 
-        public static TransformMatrix barLocalAxisToCSI = Geometry.Modify.Transpose(barLocalAxisToBHoM);
+        public const string AdapterIdName = "SAP2000_id";
+
+        /***************************************************/
     }
 }
