+++ SAP2000_Adapter/Convert/ToSAP2000/Constraint.cs
@@ -23,7 +23,7 @@
 using BH.oM.Structure.Constraints;
 using BH.oM.Structure.Elements;
 
-namespace BH.Engine.SAP2000
+namespace BH.Adapter.SAP2000
 {
     public static partial class Convert
     {
